src/requirements.txt
Django==5.1.4
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg==3.2.3
psycopg2-binary==2.9.10
python-dotenv==0.21.1
pandas==2.2.3
openpyxl==3.1.5
Pillow